+++ vector/src/main/java/im/vector/app/features/media/ImageContentRenderer.kt
@@ -85,6 +85,19 @@ class ImageContentRenderer @Inject constructor(private val activeSessionHolder: 
     }
 
     /**
+     * For url preview
+     */
+    fun render(mxcUrl: String, imageView: ImageView): Boolean {
+        val contentUrlResolver = activeSessionHolder.getActiveSession().contentUrlResolver()
+        val imageUrl = contentUrlResolver.resolveFullSize(mxcUrl) ?: return false
+
+        GlideApp.with(imageView)
+                .load(imageUrl)
+                .into(imageView)
+        return true
+    }
+
+    /**
      * For gallery
      */
     fun render(data: Data, imageView: ImageView, size: Int) {
@@ -232,7 +245,7 @@ class ImageContentRenderer @Inject constructor(private val activeSessionHolder: 
             val contentUrlResolver = activeSessionHolder.getActiveSession().contentUrlResolver()
             val resolvedUrl = when (mode) {
                 Mode.FULL_SIZE,
-                Mode.STICKER   -> resolveUrl(data)
+                Mode.STICKER -> resolveUrl(data)
                 Mode.THUMBNAIL -> contentUrlResolver.resolveThumbnail(data.url, size.width, size.height, ContentUrlResolver.ThumbnailMethod.SCALE)
             }
             // Fallback to base url
@@ -300,7 +313,7 @@ class ImageContentRenderer @Inject constructor(private val activeSessionHolder: 
                     finalHeight = min(maxImageWidth * height / width, maxImageHeight)
                     finalWidth = finalHeight * width / height
                 }
-                Mode.STICKER   -> {
+                Mode.STICKER -> {
                     // limit on width
                     val maxWidthDp = min(dimensionConverter.dpToPx(120), maxImageWidth / 2)
                     finalWidth = min(dimensionConverter.dpToPx(width), maxWidthDp)