+++ CHANGES.md
@@ -4,6 +4,7 @@ Changes in Element 1.0.12 (2020-XX-XX)
 Features ✨:
  - Add room aliases management, and room directory visibility management in a dedicated screen (#1579, #2428)
  - Room setting: update join rules and guest access (#2442)
+ - Url preview (#481)
  - Store encrypted file in cache and cleanup decrypted file at each app start (#2512)
  - Emoji Keyboard (#2520)
 
@@ -22,6 +23,7 @@ Translations 🗣:
  -
 
 SDK API changes ⚠️:
+ - RawCacheStrategy has been moved and renamed to CacheStrategy
  - FileService: remove useless FileService.DownloadMode
 
 Build 🧱: